https://chronobank.atlassian.net/browse/MD-550 - update middleware / eth / ipfs https://chronobank.atlassian.net/browse/MD-510 - middleware - tests / eth / ipfs
